pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "caspyorm"
version = "0.1.5"
authors = [
  { name="CaspyORM Team", email="team@caspyorm.dev" },
]
description = "Um ORM moderno e eficiente para Apache Cassandra"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["cassandra", "orm", "database", "pydantic", "fastapi", "nosql"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Framework :: FastAPI",
    "Typing :: Typed",
]

dependencies = [
  "cassandra-driver>=3.28.0",
  "pydantic>=2.0.0",
  "typing-extensions>=4.0.0",
  "typer[all]>=0.9.0",
  "rich>=13.0.0",
]

[project.optional-dependencies]
# Dependências opcionais para recursos assíncronos otimizados
async = ["aiocassandra"]
fastapi = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "pydantic>=2.0.0",
]

[project.scripts]
caspy = "caspyorm_cli.main:app"

[project.urls]
Homepage = "https://github.com/caspyorm/caspyorm"
Documentation = "https://caspyorm.readthedocs.io"
Repository = "https://github.com/caspyorm/caspyorm"
"Bug Tracker" = "https://github.com/caspyorm/caspyorm/issues"
"Source Code" = "https://github.com/caspyorm/caspyorm"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
    "/pyproject.toml",
]

[tool.hatch.build.targets.wheel]
packages = ["caspyorm", "caspyorm_cli"]
package-dir = {"" = "src"}
exclude = [
    "tests/",
    "scripts/",
    "examples/",
    "docs/",
    "venv/",
    "*.egg-info/",
    ".git/",
    ".pytest_cache/",
    "__pycache__/",
    "*.pyc",
    "*.pyo",
    "*.pyd",
    ".coverage",
    "htmlcov/",
    ".mypy_cache/",
    ".ruff_cache/",
    ".pypirc",
    "coverage.xml",
    "dist/",
    ".gitignore",
    "docker-compose.yml",
    "config/",
    "migrations/",
] 